+++ PhotonBeam/AS-Builders/BlasGenerator.hpp
