+++ PhotonBeam/PhotonBeamApp.cpp
@@ -8,12 +8,12 @@
 #include <imgui/imgui_impl_win32.h>
 #include <imgui/imgui_impl_dx12.h>
 #include <imgui_helper.h>
-#include <microsoft-directx-graphics-samples/raytraceHelper.hpp>
+#include <microsoft-directx-graphics-samples/raytraceHelper.h>
 #include <microsoft-directx-graphics-samples/DirectXRaytracingHelper.h>
 
 
 #include "Shaders/RaytracingHlslCompat.h"
-#include "AS-Builders/BlasGenerator.h"
+#include "AS-Builders/BlasGenerator.hpp"
 #include "Raytracing-Utils/DXCompileShader.hpp"
 
 using Microsoft::WRL::ComPtr;
